requirements.txt
Pillow>=9.0.0
openpyxl>=3.0.0 